Fixed phpdoc issues + unnecessary cast to strings
* These errors were found by using PHPStan, I simply fixed some of
  them
* There is still ~200 errors to be fixed, the most important ones
  are missing method calls, or static method calling non-static

+++ src/Faker/Provider/lt_LT/Person.php
@@ -291,7 +291,7 @@ class Person extends \Faker\Provider\Person
      * National Personal Identity number (asmens kodas)
      * @link https://en.wikipedia.org/wiki/National_identification_number#Lithuania
      * @link https://lt.wikipedia.org/wiki/Asmens_kodas
-     * @param string [male|female]
+     * @param string $gender [male|female]
      * @param \DateTime $birthdate
      * @param string $randomNumber three integers
      * @return string on format XXXXXXXXXXX
@@ -302,7 +302,7 @@ class Person extends \Faker\Provider\Person
             $birthdate = \Faker\Provider\DateTime::dateTimeThisCentury();
         }
 
-        $genderNumber = ($gender == 'male') ? (int) 1 : (int) 0;
+        $genderNumber = ($gender == 'male') ? 1 : 0;
         $firstNumber = (int) floor($birthdate->format('Y') / 100) * 2 - 34 - $genderNumber;
 
         $datePart = $birthdate->format('ymd');
@@ -318,7 +318,7 @@ class Person extends \Faker\Provider\Person
         }
 
         $sum = self::calculateSum($partOfPerosnalCode, 2);
-        $liekana = (int) $sum % 11;
+        $liekana = $sum % 11;
 
         $lastNumber = ($liekana !== 10) ? $liekana : 0;
         return $firstNumber . $datePart . $randomDigits . $lastNumber;
